reafctor: move all paypal components into the module
