+++ apps/web/modules/paypal/runtime/components/GooglePayButton.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script lang="ts" setup>
-import type { PayPalAddToCartCallback } from '~/components/PayPal/types';
+import type { PayPalAddToCartCallback } from '../types';
 import { SfLoaderCircular } from '@storefront-ui/vue';
 
 const {
