+++ apps/web/modules/paypal/runtime/components/PayPalAPM.vue
@@ -13,7 +13,7 @@
 <script setup lang="ts">
 import { cartGetters, paymentProviderGetters } from '@plentymarkets/shop-api';
 import type { PayPalNamespace, FUNDING_SOURCE, OnApproveData, OnInitActions } from '@paypal/paypal-js';
-import type { PayPalAddToCartCallback, PaypalAPMPropsType } from '~/components/PayPal/types';
+import type { PayPalAddToCartCallback, PaypalAPMPropsType } from '../types';
 import { PayPalAlternativeFundingSourceMapper } from '~/composables';
 import { SfLoaderCircular } from '@storefront-ui/vue';
