+++ apps/web/modules/paypal/runtime/components/PayPalInvoiceDetails.vue
